+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Improve `R.mapParallelAsync` typings to allow optional `batchSize` parameter.
 
+- Change order of inputs in `R.mapPropObject`
+
 10.3.3
 
 - Fix wrong typing for `R.reject` - [Issue #779](https://github.com/selfrefactor/rambda/issues/779)